fix: 줄바꿈 문제 다시 해결

+++ Command/Rm.cs
@@ -66,7 +66,12 @@ namespace VirtualTerminal.Command
 
         public string Description(bool detail)
         {
-            return "rm - 파일이나 디렉터리 삭제\n";
+            if (detail)
+            {
+                return "rm - 파일이나 디렉터리 삭제\n";
+            }
+
+            return "rm - 파일이나 디렉터리 삭제";
         }
     }
 }
